p2_continuous-control/requirements.txt
jupyter
numpy
matplotlib
tensorflow~=1.14
pytorch==1.1.0
torchvision==0.3.0
# unityagents==0.4.0
